[inline] add "All invocations in project" option
^KTIJ-542 Fixed
^KT-CR-1851

(cherry picked from commit f25816b1d82ec66bdab3b7b72a38f74aaf73f17c)

+++ kotlin-ide/idea/src/org/jetbrains/kotlin/idea/refactoring/inline/AbstractKotlinInlineNamedDeclarationProcessor.kt
@@ -115,7 +115,7 @@ abstract class AbstractKotlinInlineNamedDeclarationProcessor<TDeclaration : KtNa
     }
 
     private fun postActions() {
-        if (deleteAfter) {
+        if (deleteAfter && isWritable) {
             declaration.deleteWithCompanion()
             postDeleteAction()
         }
